backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blis==1.3.0
cached-property==2.0.1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cymem==2.0.11
es_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.8.0/es_core_news_sm-3.8.0-py3-none-any.whl#sha256=e451a83d6df79b87e9eed0cb553f03e99e36a3bab18a7b79f0dcfd1fdf875e12
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httplib2==0.22.0
huggingface-hub==0.30.2
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
packaging==25.0
pillow==11.2.1
preshed==3.0.9
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
PyPDF2==3.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==80.1.0
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
sqlalchemy-orm==1.2.10
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cached-property==2.0.1
click==8.1.8
fastapi==0.115.12
h11==0.16.0
idna==3.10
inflection==0.5.1
mypy_extensions==1.1.0
pydantic==2.11.4
pydantic_core==2.33.2
PyPDF2==3.0.1
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlalchemy-orm==1.2.10
starlette==0.46.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2